examples/requirements.txt
google-cloud-bigquery
requests
pydantic
rabbit-bq-job-optimizer 